+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.1.0...master)
 
  * Moved context.url to context.http.url for requests/urllib3 spans (#393, #394)
+ * Added support for using route as transaction name in Django 2.2+ (#86, #396)
 
 ## v4.1.0 
 [Check the diff](https://github.com/elastic/apm-agent-python/compare/v4.0.3...v4.1.0)